+++ HISTORY.rst
@@ -28,6 +28,7 @@ Next (2018-XX-XX)
 - Remove unused animate.css package.
 - Add page loading indicator.
 - Add periodic AJAX workers to update the value of the notifications link.
+- Add notification count inside a badge on the notification link.
 
 2.8 (2018-02-27)
 ================